requirements.txt
discord.py
configly
urllib
